KciNavDrawer - change architecture (and fix overflow bug)

+++ src/main/java/com/kciray/android/commons/gui/Metric.java
@@ -6,8 +6,8 @@ import com.kciray.android.commons.sys.Global;
 
 
 public class Metric {
-    public static int dpToPx(int dp) {
+    public static int dpToPx(float dp) {
         float density = Global.getContext().getResources().getDisplayMetrics().density;
-        return Math.round((float) dp * density);
+        return Math.round(dp * density);
     }
 }
